Updated mongo to version 4 to match the current version of spring. Also removed mongo versions to be consistent with all other stores and just named the module mongo without the -v3 or -v4. Added .sts4-cache to .gitignore and removed the directory from each module it was in as it is a directory that is specific to each individual development environment.
