--- ff4j-config-yaml/.sts4-cache/classpath-data.json
@@ -1,1 +1,0 @@
-{"name":"ff4j-config-yaml","classpathEntries":[{"kind":"binary","path":"/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/jre/lib/resources.jar","sourceContainerUrl":"file:/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/src.zip","javadocContainerUrl":"https://docs.oracle.com/javase/1/docs/api/","isSystem":true,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"binary","path":"/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/jre/lib/rt.jar","sourceContainerUrl":"file:/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/src.zip","javadocContainerUrl":"https://docs.oracle.com/javase/1/docs/api/","isSystem":true,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"binary","path":"/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/jre/lib/jsse.jar","sourceContainerUrl":"file:/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/src.zip","javadocContainerUrl":"https://docs.oracle.com/javase/1/docs/api/","isSystem":true,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"binary","path":"/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/jre/lib/jce.jar","sourceContainerUrl":"file:/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/src.zip","javadocContainerUrl":"https://docs.oracle.com/javase/1/docs/api/","isSystem":true,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"binary","path":"/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/jre/lib/charsets.jar","sourceContainerUrl":"file:/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/src.zip","javadocContainerUrl":"https://docs.oracle.com/javase/1/docs/api/","isSystem":true,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"binary","path":"/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/jre/lib/jfr.jar","sourceContainerUrl":"file:/Library/Java/JavaVirtualMachines/jdk1.8.0_162.jdk/Contents/Home/src.zip","javadocContainerUrl":"https://docs.oracle.com/javase/1/docs/api/","isSystem":true,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"source","path":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-core/src/main/java","outputFolder":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-core/target/classes","javadocContainerUrl":"file:/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-core/target/site/apidocs","isSystem":false,"isOwn":false,"isTest":false,"isJavaContent":true},{"kind":"source","path":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-test/src/main/java","outputFolder":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-test/target/classes","javadocContainerUrl":"file:/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-test/target/site/apidocs","isSystem":false,"isOwn":false,"isTest":false,"isJavaContent":true},{"kind":"binary","path":"/Users/cedricklunven/.m2/repository/org/yaml/snakeyaml/1.24/snakeyaml-1.24.jar","sourceContainerUrl":"file:/Users/cedricklunven/.m2/repository/org/yaml/snakeyaml/1.24/snakeyaml-1.24-sources.jar","javadocContainerUrl":"file:/Users/cedricklunven/.m2/repository/org/yaml/snakeyaml/1.24/snakeyaml-1.24-javadoc.jar","isSystem":false,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"binary","path":"/Users/cedricklunven/.m2/repository/junit/junit/4.12/junit-4.12.jar","sourceContainerUrl":"file:/Users/cedricklunven/.m2/repository/junit/junit/4.12/junit-4.12-sources.jar","javadocContainerUrl":"file:/Users/cedricklunven/.m2/repository/junit/junit/4.12/junit-4.12-javadoc.jar","isSystem":false,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"binary","path":"/Users/cedricklunven/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3.jar","sourceContainerUrl":"file:/Users/cedricklunven/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3-sources.jar","javadocContainerUrl":"file:/Users/cedricklunven/.m2/repository/org/hamcrest/hamcrest-core/1.3/hamcrest-core-1.3-javadoc.jar","isSystem":false,"isOwn":false,"isTest":false,"isJavaContent":false},{"kind":"source","path":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/src/main/java","outputFolder":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/classes","javadocContainerUrl":"file:/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/site/apidocs","isSystem":false,"isOwn":true,"isTest":false,"isJavaContent":true},{"kind":"source","path":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/src/test/java","outputFolder":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/test-classes","javadocContainerUrl":"file:/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/site/apidocs","isSystem":false,"isOwn":true,"isTest":true,"isJavaContent":true},{"kind":"source","path":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/src/main/resources","outputFolder":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/classes","isSystem":false,"isOwn":true,"isTest":false,"isJavaContent":false},{"kind":"source","path":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/src/test/resources","outputFolder":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/test-classes","isSystem":false,"isOwn":true,"isTest":true,"isJavaContent":false},{"kind":"source","path":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/src/test/java","outputFolder":"/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/test-classes","javadocContainerUrl":"file:/Users/cedricklunven/dev/WORKSPACES/FF4J/ff4j/ff4j-config-yaml/target/site/apidocs","isSystem":false,"isOwn":true,"isTest":false,"isJavaContent":false}]}
