--- ff4j-store-mongodb-v3/.gitignore
@@ -1,1 +1,0 @@
-/bin
