+++ ff4j-store-mongodb/.gitignore
@@ -1,0 +1,2 @@
+/bin
+.sts4-cache
