+++ ff4j-store-mongodb/src/test/resources/ff4j.xml
