+++ pom.xml
@@ -184,7 +184,7 @@
         <module>ff4j-store-hbase</module>
         <module>ff4j-store-ignite</module>
         <module>ff4j-store-jcache</module>
-        <module>ff4j-store-mongodb-v3</module>
+        <module>ff4j-store-mongodb</module>
         <module>ff4j-store-neo4j</module>
         <module>ff4j-store-redis</module>
         <module>ff4j-store-springjdbc</module>
